requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
pillow==10.4.0
pytz==2018.5
sqlparse==0.3.1
typing_extensions==4.12.2
pyscopg